Refactoring Shaka Agent
